# docs/api/openapi.yaml
...
components:
  schemas:
    ApiKeyIn:
      properties:
        base_url:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Base Url
        label:
          default: ''
          maxLength: 64
          title: Label
          type: string
        provider:
          enum:
          - deepseek
          - qwen
          - glm
          - moonshot
          - doubao
          - siliconflow
          - minimax
          - hunyuan
          - openai_compatible
          - openai
          - anthropic
          - gemini
          title: Provider
          type: string
        secret:
          maxLength: 512
          minLength: 8
          title: Secret
          type: string
      required:
      - provider
      - secret
      title: ApiKeyIn
      type: object
    ApiKeyOut:
      properties:
        base_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Base Url
        created_at:
          format: date-time
          title: Created At
          type: string
        enabled:
          title: Enabled
          type: boolean
        id:
          title: Id
          type: string
        label:
          title: Label
          type: string
        last_test_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Last Test Json
        provider:
          title: Provider
          type: string
        secret_mask:
          title: Secret Mask
          type: string
        updated_at:
          format: date-time
          title: Updated At
          type: string
      required:
      - id
      - provider
      - base_url
      - label
      - secret_mask
      - enabled
      - created_at
      - updated_at
      title: ApiKeyOut
      type: object
    ApiKeyProbeIn:
      description: 不落库的试连：用户刚粘贴 Key 时先试，避免存一堆废 Key。
      properties:
        base_url:
          anyOf:
          - maxLength: 255
            type: string
          - type: 'null'
          title: Base Url
        model:
          anyOf:
          - maxLength: 64
            type: string
          - type: 'null'
          title: Model
        provider:
          enum:
          - deepseek
          - qwen
          - glm
          - moonshot
          - doubao
          - siliconflow
          - minimax
          - hunyuan
          - openai_compatible
          - openai
          - anthropic
          - gemini
          title: Provider
          type: string
        secret:
          maxLength: 512
          minLength: 8
          title: Secret
          type: string
      required:
      - provider
      - secret
      title: ApiKeyProbeIn
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    KvIn:
      properties:
        value:
          title: Value
          type: string
      required:
      - value
      title: KvIn
      type: object
...
    ModelProfileIn:
      properties:
        name:
          maxLength: 64
          minLength: 1
          title: Name
          type: string
        text_key_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Text Key Id
        text_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Text Model
        vision_key_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Vision Key Id
        vision_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Vision Model
      required:
      - name
      title: ModelProfileIn
      type: object
    ModelProfileOut:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        id:
          title: Id
          type: string
        name:
          title: Name
          type: string
        text_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Text Model
        text_provider_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Text Provider Id
        updated_at:
          format: date-time
          title: Updated At
          type: string
        vision_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Vision Model
        vision_provider_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Vision Provider Id
      required:
      - id
      - name
      - text_provider_id
      - text_model
      - vision_provider_id
      - vision_model
      - created_at
      - updated_at
      title: ModelProfileOut
      type: object
...
    PasswordChangeIn:
      description: '改口令输入。``old_password`` 只校验非空 —— 强度是**当前**口令的事，

        强制它满足现行强度规则会让「弱口令时代设的旧口令」永远改不了。'
      properties:
        new_password:
          maxLength: 1024
          minLength: 1
          title: New Password
          type: string
        old_password:
          maxLength: 1024
          minLength: 1
          title: Old Password
          type: string
      required:
      - old_password
      - new_password
      title: PasswordChangeIn
      type: object
...
    SessionOut:
      description: '登录/创建口令成功后下发的会话令牌。


        令牌是 HMAC 签名（不是随机串），因此服务端无需会话表；它的有效期由

        签发时的 TTL 决定，改口令会让它立刻失效。**只在响应里出现，不写日志。**'
      properties:
        expires_at:
          format: date-time
          title: Expires At
          type: string
        session_token:
          title: Session Token
          type: string
      required:
      - session_token
      - expires_at
      title: SessionOut
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: '学习工作台 Strayt 服务端。


    单用户免账号：所有 `/api/v1` 端点需 `Authorization: Bearer <访问令牌>`。

    唯一豁免是 `/health`（存活探针）。


    错误体统一为 `{ "error": { "code", "message", "detail?" } }`，`code` 是稳定契约，客户端据此分支。'
  title: Strayt 学习工作台
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/auth/password:
    post:
      description: '改口令。鉴权由 ``router2`` 在 ``main.py`` 挂载时用 ``AuthDep`` 施加。


        改完立刻返回**新**会话令牌：签名密钥由口令哈希派生，换口令等于换密钥，

        客户端手上那个旧令牌当场作废。不回传的话用户会被自己踢下线，得重新输一遍。'
      operationId: change_password_api_v1_auth_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Change Password
      tags:
      - 访问口令登录
...
  /api/v1/settings/api-keys:
    get:
      operationId: list_keys_api_v1_settings_api_keys_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ApiKeyOut'
                title: Response List Keys Api V1 Settings Api Keys Get
                type: array
          description: Successful Response
      summary: Key 列表（只含掩码）
      tags:
      - 设置
    post:
      operationId: create_key_api_v1_settings_api_keys_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 添加 Key
      tags:
      - 设置
  /api/v1/settings/api-keys/probe:
    post:
      operationId: probe_key_api_v1_settings_api_keys_probe_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyProbeIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Probe Key Api V1 Settings Api Keys Probe Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 试连（不落库，≤10s）
      tags:
      - 设置
  /api/v1/settings/api-keys/{key_id}:
    delete:
      operationId: delete_key_api_v1_settings_api_keys__key_id__delete
      parameters:
      - in: path
        name: key_id
        required: true
        schema:
          title: Key Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 删除 Key
      tags:
      - 设置
  /api/v1/settings/api-keys/{key_id}/test:
    post:
      description: F6 硬要求：10s 内返回成功/失败 + 原因（Key 错误 / 网络不通 / 额度不足）。
      operationId: test_key_api_v1_settings_api_keys__key_id__test_post
      parameters:
      - in: path
        name: key_id
        required: true
        schema:
          title: Key Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 连通性测试（≤10s，F6）
      tags:
      - 设置
  /api/v1/settings/kv/{key}:
    get:
      operationId: get_kv_api_v1_settings_kv__key__get
      parameters:
      - in: path
        name: key
        required: true
        schema:
          title: Key
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Get Kv Api V1 Settings Kv  Key  Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 读 KV
      tags:
      - 设置
    put:
      operationId: put_kv_api_v1_settings_kv__key__put
      parameters:
      - in: path
        name: key
        required: true
        schema:
          title: Key
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KvIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Put Kv Api V1 Settings Kv  Key  Put
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 写 KV
      tags:
      - 设置
  /api/v1/settings/model-profiles:
    get:
      operationId: list_profiles_api_v1_settings_model_profiles_get
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ModelProfileOut'
                title: Response List Profiles Api V1 Settings Model Profiles Get
                type: array
          description: Successful Response
      summary: 模型档案列表
      tags:
      - 设置
    post:
      operationId: create_profile_api_v1_settings_model_profiles_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelProfileIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelProfileOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 新建档案
      tags:
      - 设置
  /api/v1/settings/model-profiles/{profile_id}:
    delete:
      operationId: delete_profile_api_v1_settings_model_profiles__profile_id__delete
      parameters:
      - in: path
        name: profile_id
        required: true
        schema:
          title: Profile Id
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 删除档案
      tags:
      - 设置
...